Matej: * controllers = (auth.php, email_types.php, event_categories.php, events.php, payments.php, posts.php, studies.php) boli okomentovane a v co najvacsej miere zdebugovane a nastavene ako maju byt * models = inserter, selecter, updater boli miestami doplnene resp upravene(fixnute) * views = gridy na degrees, email_types, event_categories, events, payments boli translatnute do SK jazyku(stlpce) + poschovavane co nema byt vidiet IDcka, + ak nebolo nastavene dropdowny * events ako komponent by mal byt kompletne pojazdny skrz naskrz vsetky metody * okomentovane userdata, general_helper.php * posts ako komponent by mal v co najvacsej miere ficat, avsak mensie bugy stale niesu vylucene -------------------------------------------------------------------------------------------------- ****Celkovo by to cele hento malo byt funkcne avsak na neskoru nocnu sry rannu hodinu sa mozu vyskytnut este mensie bugy ospatostou autora, bugy hlaste mne, za pochopenie dakuje :-) ****

+++ trunk/application/views/posts_detail_view.php
@@ -4,26 +4,30 @@
 ?>
 
 <div class="post_title">
-    <?= $obj[0]->post_title ?>
+    <?= $obj->post_title ?>
 </div>
 
-<div id="post_options">
-    <?= anchor('posts/delete/'.$post_id, $this->lang->line('delete_item')); ?>
-    <?= anchor('posts/edit/'.$post_id, $this->lang->line('edit_item')); ?>
-</div>
+<?php if( $this->userdata->is_admin() ): ?>
+    <div id="post_options">
+        <?= anchor('posts/delete/'.$post_id, $this->lang->line('delete_item')); ?>
+        <?= anchor('posts/edit/'.$post_id, $this->lang->line('edit_item')); ?>
+    </div>
+<?php endif; ?>
 
 <div class="post_content">
-    <?= parse_bbcode($obj[0]->post_content) ?>  
+    <?= parse_bbcode($obj->post_content) ?>  
 </div>
 
 <div class="post_add_info">
     <span><?= $this->lang->line('created_by'); ?>:</span>
-    <?= anchor('users/detail/'.$obj[0]->post_author_id, $obj[0]->author_name.' '.$obj[0]->author_surname); ?> ,
-    <?= $obj[0]->post_date ?> 
+    <?= anchor('users/detail/'.$obj->post_author_id, $obj->author_name.' '.$obj->author_surname); ?>, 
+    <?= datetime($obj->post_date, FALSE).' '. time_withou_seconds(datetime($obj->post_date, TRUE)) ?> 
 </div>
 
-<div class="post_modifie_info">
-    <span><?= $this->lang->line('last_update'); ?></span>
-    <?= anchor('users/detail/'.$obj[0]->post_modifie_author_id, $obj[0]->modifie_name.' '.$obj[0]->modifie_surname) ?> ,
-    <?= $obj[0]->post_modifie_date ?> 
-</div>
+<?php if($obj->post_modifie_date == TRUE && $obj->post_modifie_author_id == TRUE): ?>
+    <div class="post_modifie_info">
+        <span><?= $this->lang->line('last_update'); ?></span>
+        <?= anchor('users/detail/'.$obj->post_modifie_author_id, $obj->modifie_name.' '.$obj->modifie_surname) ?>,
+        <?= datetime($obj->post_modifie_date, FALSE).' '. time_withou_seconds(datetime($obj->post_modifie_date, TRUE)) ?> 
+    </div>
+<?php endif; ?>
